% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitsFinder.R
\name{autoSplits}
\alias{autoSplits}
\title{splitsFinder}
\usage{
autoSplits(data, w, span)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader()}

\item{w}{(numeric) w should be << than half of the width of data. small values gets tiny bumps}

\item{span}{(numeric) span 0 to 1. larger makes bump go away
Use columns = 'all' to plot all the metrics}
}
\value{
A list of dataframes, one for each split
}
\description{
Function to automatically detect splits
}
\details{
The function accepts a data frame created with dataLoader()
}
\examples{
# gpx <- dataLoader("path_to_activity.tcx")
# generateMap(gpx)
# autoSp<-autoSplits(gpx, 1, 1.1)
}
