% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSplits.R
\name{compareSplits}
\alias{compareSplits}
\title{compareSplits}
\usage{
compareSplits(data, showMe = NULL, ftp = NULL, ftpType = c("power",
  "pace", "HR")[1])
}
\arguments{
\item{data}{an activity dataframe or list of dataframes (splits)}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{ftp}{(numeric) Functional threshold power if bike workout, pace [min/km] if run or swim, lactate heart rate
threshold if using HR}

\item{ftpType}{(character) One of 'power', 'pace' or 'HR', to describe what kind of FTP
was provided}
}
\value{
A dataframe with a summary of comparisons between metrics for the activity.
}
\description{
Function to generate a summary of the activity data
}
\details{
The function summarise the activity data in a dataframe. If a valid FTPO and FTP type combination is provided
the function also computes TSS and IF (see Joe Friel blog for more details on these metrics)
}
\examples{
gpx <- intervalActivity
compareSplits(gpx, ftp = 4.1, ftpType = "pace")
sp<-autoSplits(gpx)
compareSplits(sp, ftp = 4.1, ftpType = "pace")
}
