% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupling.R
\name{decoupling}
\alias{decoupling}
\title{decoupling}
\usage{
decoupling(data, variable, from, to)
}
\arguments{
\item{data}{an activity dataframe}

\item{variable}{(character) The metrics to compare in the decaopling analysis against heart rate.
Must be the same name as the column in the dataframe.}

\item{from}{(numeric) When to start the decoupling analysis}

\item{to}{(numeric) When to stop the decoupling analysis}
}
\value{
A list of dataframes with a summary of comparisons between heart rate (1st and 2nd half, compareHR),
variable (1st and 2nd half, compareVar), and heart rate against variable (overall, compareSlopes).
It also returns the decoupling plot (p).
}
\description{
Function to analyse activity decoupling
}
\details{
The concept of decoupling is discussed in Joe Friel blog. In summary, this function allows to visualise and 
quantify if an activity (ideally a long steady effort in zone 2) shows decoupling or not, that is if the  heart rate stays
constant but the pace/power decrease or if the the pace/power stay constant, but the heart rate increases.
High decoupling is symptomatic of poor aerobic fitness.
}
\examples{
gpx <- evenActivity
decoplingRes <- decoupling(data = gpx, variable = "Pace", from = 10, to = 60)
decoplingRes$compareHR
decoplingRes$compareVar
decoplingRes$compareSlopes
print(decoplingRes$p)
}
