% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evenActivity}
\alias{evenActivity}
\title{Example of activity not containing intervals.}
\format{A data frame with 4324 rows and 11 variables:
\describe{
  \item{Time}{The recorded time}
  \item{Position}{a}
  \item{AltitudeMeters}{a}
  \item{DistanceMeters}{a}
  \item{HeartRateBpm}{a}
  \item{Extensions}{a}
  \item{lat}{a}
  \item{lon}{a}
  \item{AltitudeMetersDiff}{a}
  \item{Pace}{a}
  \item{Speed}{a}
}}
\usage{
evenActivity
}
\description{
A dataset resulting from loading a tcx file
of an activity without intervals
}
\keyword{datasets}
