% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMap.R
\name{generateMap}
\alias{generateMap}
\title{generateMap}
\usage{
generateMap(data)
}
\arguments{
\item{data}{an activity dataframe or a list of activity dataframes (splits)}
}
\value{
A leaflet object to view as a map
}
\description{
Function to generate a map of the activity
}
\details{
If the activity dataframe(s) contains latitude and longitude informations, then this
function will plot the activity trackpoints on a map
}
\examples{
gpx <- evenActivity
generateMap(gpx)
sp <- autoSplits(gpx)
generateMap(sp)
}
