% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitsValues.R
\name{getSplitsValues}
\alias{getSplitsValues}
\title{getSplitsValues}
\usage{
getSplitsValues(gpx, sp, xVariable)
}
\arguments{
\item{gpx}{(data frame) a dataframe generated with gpxAnalyser::dataLoader()}

\item{sp}{(list) the list of dataframe representing the splits}

\item{xVariable}{(character) the variable to put on the x axis. One of 'DistanceMeters'
or 'Time'}
}
\value{
A numeric vector
}
\description{
Function to detSplitsValues
}
\details{
the function gives the splits intital time or distance, needed to plot the splits on the metrics plot
}
\examples{
gpx <- intervalActivity
sp <- autoSplits(gpx)
splitsValues <- getSplitsValues(gpx, sp, "Time")
}
