% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothData.R
\name{gpxSmooth}
\alias{gpxSmooth}
\title{gpxSmooth}
\usage{
gpxSmooth(data, variable = NULL, k, endrule = c("median", "keep",
  "constant")[1], algorithm = NULL, print.level = 0)
}
\arguments{
\item{data}{(dataframe) an activity dataframe}

\item{variable}{(character) the column name of the variable to smooth. If NULL, all
variable except 'Time', 'Distance', 'LatitudeDegrees' and 'LongitudeDegrees' will be
smoothed}
}
\value{
An activity dataframe with smoothed data
}
\description{
Function to smooth  variable. The function uses the function
runmed() to smooth the data. See its documentation for
arguents and tier meaning
}
\details{
The function uses the provided split values and split type to divide the activity dataframe in a list of dataframes,
each one corresponding to a split starting and finishing at the specified times/distances.
}
\examples{
gpx <- evenActivity
smoothedData <- gpxSmooth(gpx, "Pace", k = 39)
}
