% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervalActivity}
\alias{intervalActivity}
\title{Example of activity containing intervals.}
\format{A data frame with 3730 rows and 11 variables:
\describe{
  \item{Time}{The recorded time}
  \item{Position}{a}
  \item{AltitudeMeters}{a}
  \item{DistanceMeters}{a}
  \item{HeartRateBpm}{a}
  \item{Extensions}{a}
  \item{lat}{a}
  \item{lon}{a}
  \item{AltitudeMetersDiff}{a}
  \item{Pace}{a}
  \item{Speed}{a}
  }}
\usage{
intervalActivity
}
\description{
A dataset resulting from loading a tcx file
of an activity with intervals
}
\keyword{datasets}
