% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purgeSplits.R
\name{purgeSplits}
\alias{purgeSplits}
\title{purgeSplits}
\usage{
purgeSplits(data, splitsToEliminate = NULL)
}
\arguments{
\item{data}{a list of dataframes generated with gpxAnalyser::createSplits()}

\item{splitsToEliminate}{(numeric vector) the indexes of the splits
to eliminate from the map view. Default is NULL (no split is eliminated).}
}
\value{
A list of data frames
}
\description{
Function to a eliminate splits
}
\details{
The function accepts a list of dataframes createdwith createSplits() or dataLoader()
}
\examples{
gpx <- evenActivity
sp <- createSplits(gpx, 2000, type = "everyKm")
purgedSplits <- purgeSplits(sp, splitsToEliminate = c(1))
}
