% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpikes.R
\name{removeSpikes}
\alias{removeSpikes}
\title{removeSpikes}
\usage{
removeSpikes(data, tolLevel = 0.1, span = 0.5)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader() or the
list of dataframes generated with gpxAnalyser::createSplits()}

\item{tolLevel}{(numeric) the tolerance level}

\item{span}{(numeric) the span value for the smooth function.
Big span means big smooth}
}
\value{
The given object without the spikes
}
\description{
Function to remove spikes in metrics
}
\details{
The function accepts a list of dataframes or a data frame created
with createSplits() or dataLoader()
}
\examples{
# gpx <- dataLoader("path_to_activity.tcx")
# generateMap(gpx)
# sp<-createSplits(gpx, 2000, type = "everyKm")
# sp<-removeSpikes(sp)
}
