% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryGpx.R
\name{summaryGpx}
\alias{summaryGpx}
\title{summaryGpx}
\usage{
summaryGpx(data)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader() or the
list of dataframes generated with gpxAnalyser::createSplits()}
}
\value{
A dataframe summarising activity metrics
}
\description{
Function to a map of the activity
}
\details{
The function accepts a list of dataframes or a data frame created
with createSplits() or dataLoader()
}
\examples{
# gpx <- dataLoader("path_to_activity.tcx")
# sp<-createSplits(gpx, 2000, type = "everyKm")
# overview<-summaryGpx(gpx)
# overview<-summaryGpx(sp)
}
