% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSplits.R
\name{autoSplits}
\alias{autoSplits}
\title{splitsFinder}
\usage{
autoSplits(data)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader()}
}
\value{
A list of dataframes, one for each split
}
\description{
Function to automatically detect splits. If the splits were recorded by the Garmin device, the function
will find and return those
}
\details{
The function accepts a data frame created with dataLoader()
}
\examples{
gpx <- intervalActivity
autoSp <- autoSplits(gpx) #to use Garmin splits
}
