% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSplits.R
\name{autoSplits}
\alias{autoSplits}
\title{autoSplits}
\usage{
autoSplits(data)
}
\arguments{
\item{data}{an activity dataframe}
}
\value{
A list of dataframes, one for each split
}
\description{
Function to automatically detect splits recorded in the activity.
}
\details{
Sometimes intervals are recorded directly in the activity, for instance when pressing the Lap button on a
Garmin device. This function will identify such splits, if any is present. The splits are identified as
repeated rows in the data frame.
}
\examples{
gpx <- intervalActivity
autoSp <- autoSplits(gpx)
}
