% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSplits.R
\name{autoSplits}
\alias{autoSplits}
\title{splitsFinder}
\usage{
autoSplits(data, span = NULL, yVariable = "DistanceMeters")
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader()}

\item{span}{(numeric) a numeric value. The smaller the values, the more splits you get (normally)}

\item{yVariable}{(numeric) the variable (column name) to put on the y axis}
}
\value{
A list of dataframes, one for each split
}
\description{
Function to automatically detect splits. If the splits were recorded by the Garmin device, the function
will find and return those
}
\details{
The function accepts a data frame created with dataLoader()
}
\examples{
gpx <- intervalActivity
autoSp <- autoSplits(gpx) #to use Garmin splits
# If the activity had no Garmin splits:
# autoSp <- autoSplits(gpx, span = 25) 
}
