% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSplits.R
\name{compareSplits}
\alias{compareSplits}
\title{compareSplits}
\usage{
compareSplits(data)
}
\arguments{
\item{data}{a list of dataframes generated with gpxAnalyser::createSplits()}
}
\value{
A dataframe with a summary of comparisons between metrics for the splits
}
\description{
Function to compare activity splits
}
\details{
The function accepts a list of dataframes created with createSplits()
}
\examples{
# gpx <- dataLoader("path_to_activity.tcx")
# sp<-createSplits(gpx, 2000, type = "everyKm")
# compare<-createSplits(sp)
}
