% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSplits.R
\name{compareSplits}
\alias{compareSplits}
\title{compareSplits}
\usage{
compareSplits(data, showMe, ftp = NULL, ftpType = c("power", "pace",
  "HR")[1])
}
\arguments{
\item{data}{a list of dataframes generated with gpxAnalyser::createSplits()}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{ftp}{(numeric) Functional threshold power if bike workout, pace [min/km] if run, functional threshold if using HR}

\item{ftpType}{(character) One os 'power', 'pace' or 'HR', to describe what kind of FTP
was provided}
}
\value{
A dataframe with a summary of comparisons between metrics for the splits
}
\description{
Function to compare activity splits
}
\details{
The function accepts a list of dataframes created with createSplits()
}
\examples{
gpx <- intervalActivity
compareSplits(gpx, showMe = colnames(gpx))
sp<-autoSplits(gpx)
compareSplits(sp, showMe = colnames(sp[[1]]))
}
