% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPlots.R
\name{doPlots}
\alias{doPlots}
\title{doPlots}
\usage{
doPlots(
  data,
  xVariable = c("Distance", "Time")[2],
  showMe = NULL,
  doFacet = TRUE,
  sp = NULL
)
}
\arguments{
\item{data}{an activity dataframe}

\item{xVariable}{(character) the variable to put on the x axis. One of 'DistanceMeters'
or 'Time'}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{doFacet}{(logical) TRUE to plot variables separetly, FALSE to plot
them overlayed}

\item{sp}{(list) the list of frame of splits, if plotting splits}
}
\value{
A ggplot object
}
\description{
Function to plot activity metrics
}
\details{
The function accepts a list of dataframes or a data frame created
with createSplits() or dataLoader()
}
\examples{
gpx <- evenActivity
doPlots(gpx)
doPlots(gpx, doFacet = TRUE)
sp<-autoSplits(gpx)
doPlots(gpx, sp = sp)
}
