% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPlots.R
\name{doPlots}
\alias{doPlots}
\title{doPlots}
\usage{
doPlots(data, xVariable = c("DistanceMeters", "Time")[2], showMe,
  doFacet = TRUE)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader() or the
list of dataframes generated with gpxAnalyser::createSplits()}

\item{xVariable}{(character) the variable to put on the x axis. One of 'DistanceMeters'
or 'Time'}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{doFacet}{(logical) TRUE to plot variables separetly, FALSE to plot
them overlayed}
}
\value{
A ggplot object
}
\description{
Function to plot activity metrics
}
\details{
The function accepts a list of dataframes or a data frame created
with createSplits() or dataLoader()
}
\examples{
gpx <- evenActivity
doPlots(gpx, showMe = colnames(gpx))
doPlots(gpx, doFacet = TRUE, showMe = colnames(gpx))
sp<-createSplits(gpx, 2000, type = "everyKm")
doPlots(sp, showMe = colnames(sp[[1]]))
doPlots(sp, doFacet = TRUE, showMe = colnames(sp[[1]]))
}
