% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{evenActivity}
\alias{evenActivity}
\title{Example of activity not containing intervals.}
\format{A data frame with 5192 rows and 10 variables:
\describe{
  \item{Time}{The recorded time}
  \item{Elevation}{The altitude above sea level}
  \item{Distance}{The distance done [meters]}
  \item{Heart rate}{The hear rate}
  \item{LatitudeDegrees}{The calculated latitude}
  \item{LongitudeDegrees}{The calculated longitude}
  \item{Elevation gain}{The elevation gained/lost}
  \item{Pace}{The pace [min/km]}
  \item{Speed}{The speed [km/h]}
  \item{GAP}{The gradient adjusted pace [min/Km]}
}}
\usage{
evenActivity
}
\description{
A dataset resulting from loading a tcx file
of an activity without intervals, i.e. a steady effort activity.
}
\keyword{datasets}
