% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMap.R
\name{generateMap}
\alias{generateMap}
\title{generateMap}
\usage{
generateMap(data)
}
\arguments{
\item{data}{a dataframe generated with gpxAnalyser::dataLoader() or the
list of dataframes generated with gpxAnalyser::createSplits()}
}
\value{
A leaflet object to view as a map
}
\description{
Function to a map of the activity
}
\details{
The function accepts a list of dataframes or a data frame created
with createSplits() or dataLoader()
}
\examples{
gpx <- intervalActivity
generateMap(gpx)
sp<-createSplits(gpx, 2000, type = "everyKm")
generateMap(sp)
}
