% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervalActivity}
\alias{intervalActivity}
\title{Example of activity containing intervals.}
\format{A data frame with 4051 rows and 9 variables:
\describe{
  \item{Time}{The recorded time}
  \item{AltitudeMeters}{The altitude above sea level}
  \item{DistanceMeters}{The distance done [meters]}
  \item{HeartRateBpm}{The hear rate}
  \item{LatitudeDegrees}{The calculated latitude}
  \item{LongitudeDegrees}{The calculated longitude}
  \item{AltitudeMetersDiff}{The elevation gained/lost}
  \item{Pace}{The pace [min/km]}
  \item{Speed}{The speed [km/h]}
  }}
\usage{
intervalActivity
}
\description{
A dataset resulting from loading a tcx file
of an activity with intervals
}
\keyword{datasets}
