% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompareSplits.R
\name{plotCompareSplit}
\alias{plotCompareSplit}
\title{plotCompareSplits}
\usage{
plotCompareSplit(data, doFacet = TRUE)
}
\arguments{
\item{data}{(list) the list of dataframes generated with
gpxAnalyser::createSplits()}

\item{doFacet}{(logical) TRUE to plot variables separetly, FALSE to plot
them overlayed}
}
\value{
A ggplot object
}
\description{
Function to plot activity metrics comparison between splits
}
\details{
The function accepts a list of dataframes created
with createSplits()
}
\examples{
# gpx <- dataLoader("path_to_activity.tcx")
# generateMap(gpx)
# sp<-createSplits(gpx, 2000, type = "everyKm")
# plotCompareSplits(sp)
}
