% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompareSplits.R
\name{plotCompareSplit}
\alias{plotCompareSplit}
\title{plotCompareSplits}
\usage{
plotCompareSplit(data, doFacet = TRUE, type = c("raw", "norm", "perc")[1],
  xVariable = "Time", showMe = NULL, ftp = NULL, ftpType = c("power",
  "pace", "HR")[1])
}
\arguments{
\item{data}{(list) the list of dataframes for the activity splits}

\item{doFacet}{(logical) TRUE to plot variables separetly, FALSE to plot
them overlayed}

\item{type}{(character) One of "raw", "norm" or "perc" to plot differently
normalised valued.
\itemize{
\item 'raw' is the raw data
\item 'norm' is the data normalised in the rage 0 to 100
\item 'perc' is the variation in percentage from the mean value of all the intervals
}
If using 'raw' or 'perc' it is suggested to set doFacet = TRUE}

\item{xVariable}{(character) the variable to put on the x axis. One of 'DistanceMeters'
or 'Time'}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{ftp}{(numeric) Functional threshold power if bike workout, pace [min/km] if run, functional threshold if using HR}

\item{ftpType}{(character) One os 'power', 'pace' or 'HR', to describe what kind of FTP
was provided}
}
\value{
A ggplot object
}
\description{
Function to plot activity metrics comparison between splits
}
\details{
The function plots the variation of every metric over interval index, It is useful to see if all interval were performed
at the desired target intensity
}
\examples{
gpx <- intervalActivity
sp <- autoSplits(gpx)
plotCompareSplit(sp, doFacet=TRUE, type = "raw")
plotCompareSplit(sp, doFacet=TRUE, type = "norm")
plotCompareSplit(sp, doFacet=FALSE, type = "norm")
plotCompareSplit(sp, doFacet=TRUE, type = "perc")
}
