% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCompareSplits.R
\name{plotCompareSplit}
\alias{plotCompareSplit}
\title{plotCompareSplits}
\usage{
plotCompareSplit(data, doFacet = TRUE, type = c("raw", "norm", "perc")[2],
  xVariable = "Time", showMe, ftp = NULL)
}
\arguments{
\item{data}{(list) the list of dataframes generated with
gpxAnalyser::createSplits()}

\item{doFacet}{(logical) TRUE to plot variables separetly, FALSE to plot
them overlayed}

\item{type}{(character) One of "raw", "norm" or "perc" to plot differently
normalised valued. Raw is the raw data. Norm is the data normalised in the rage 0 to 100,
perc is the variation in percentage from the mean value of all the intervals. If using 'raw'
or 'perc' it is suggested to set doFacet = TRUE}

\item{xVariable}{(character) the variable to put on the x axis. One of 'DistanceMeters'
or 'Time'}

\item{showMe}{(character) The metrics to include in the plot. Must be the same name as the column in the dataframe.}

\item{ftp}{(numeric) Functional threshold power if bike workout, pace [min/km] if run, functional threshold if using HR}
}
\value{
A ggplot object
}
\description{
Function to plot activity metrics comparison between splits
}
\details{
The function accepts a list of dataframes created
with createSplits()
}
\examples{
gpx <- intervalActivity
sp <- autoSplits(gpx)
plotCompareSplit(sp, doFacet=TRUE, type = "raw", showMe = colnames(sp[[1]]))
plotCompareSplit(sp, doFacet=TRUE, type = "norm", showMe = colnames(sp[[1]]))
plotCompareSplit(sp, doFacet=FALSE, type = "norm", showMe = colnames(sp[[1]]))
plotCompareSplit(sp, doFacet=TRUE, type = "perc", showMe = colnames(sp[[1]]))
}
