% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purgeSplits.R
\name{purgeSplits}
\alias{purgeSplits}
\title{purgeSplits}
\usage{
purgeSplits(data, splitsToEliminate = NULL)
}
\arguments{
\item{data}{a list of dataframes for the activity splits}

\item{splitsToEliminate}{(numeric vector) the indexes of the splits
to eliminate from the map view. Default is NULL (no split is eliminated).}
}
\value{
A list of data frames
}
\description{
Function to a eliminate splits
}
\details{
This function allows to eliminate splits from the views and the summary tables. This can be useful when an activity includes
rest intervals, and we are interest in metrics that would be biased by the low intensity rest interval (i.e. total intervals
time for a 5X1Km session).
}
\examples{
gpx <- intervalActivity
sp <- autoSplits(gpx)
purgedSplits <- purgeSplits(sp, splitsToEliminate = c(1,3,5,7,9,11,12))
}
